Reintroducing 9/port
- Removing obsolete syscalls
- Adding nsec
- Fixing types
- Fixing portclock ms2tk and removing unneeded macros

+++ sys/src/9/port/devcons.c
@@ -776,7 +776,7 @@ static int32_t
 consread(Chan *c, void *buf, int32_t n, int64_t off)
 {
 	Proc *up = externup();
-	uint32_t l;
+	uint64_t l;
 	Mach *mp;
 	char *b, *bp, ch, *s, *e;
 	char tmp[512];		/* Qswap is 381 bytes at clu */
